+++ src/main/java/model/piece/Cannon.java
@@ -1,8 +1,14 @@
-package model;
+package model.piece;
 
 import java.util.HashSet;
 import java.util.Set;
 import java.util.stream.Collectors;
+import model.Color;
+import model.Direction;
+import model.OccupiedPositions;
+import model.PieceIdentity;
+import model.PieceType;
+import model.Position;
 
 public class Cannon extends Piece {
