+++ src/main/java/model/piece/Guard.java
@@ -1,13 +1,19 @@
-package model;
+package model.piece;
 
 import java.util.Arrays;
 import java.util.Set;
 import java.util.stream.Collectors;
+import model.Color;
+import model.Direction;
+import model.OccupiedPositions;
+import model.PieceIdentity;
+import model.PieceType;
+import model.Position;
 
-public class King extends Piece {
+public class Guard extends Piece {
 
-    protected King(Color color) {
-        super(new PieceIdentity(color, PieceType.KING));
+    public Guard(Color color) {
+        super(new PieceIdentity(color, PieceType.GUARD));
     }
 
     @Override
