+++ src/main/java/model/piece/King.java
@@ -1,26 +1,28 @@
-package model;
+package model.piece;
 
+import java.util.Arrays;
 import java.util.Set;
 import java.util.stream.Collectors;
+import model.Color;
+import model.Direction;
+import model.OccupiedPositions;
+import model.PieceIdentity;
+import model.PieceType;
+import model.Position;
 
-public class Soldier extends Piece {
+public class King extends Piece {
 
-    protected Soldier(Color color) {
-        super(new PieceIdentity(color, PieceType.SOLDIER));
+    public King(Color color) {
+        super(new PieceIdentity(color, PieceType.KING));
     }
 
     @Override
     public Set<Position> calculateMovablePositions(Position startPosition, OccupiedPositions occupiedPositions) {
-        return Direction.getStraightDirection().stream()
-                .filter(this::isNotBack)
+        return Arrays.stream(Direction.values())
                 .filter(startPosition::canMove)
                 .map(startPosition::move)
                 .filter(destination -> destinationIsNotSameColor(occupiedPositions, destination))
                 .collect(Collectors.toSet());
-    }
-
-    private boolean isNotBack(Direction direction) {
-        return direction != Direction.calculateBackDirection(identity().getColor());
     }
 
     private boolean destinationIsNotSameColor(OccupiedPositions occupiedPositions, Position destination) {
