+++ src/test/java/model/BoardTest.java
@@ -3,8 +3,7 @@ package model;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import java.util.Map;
-import java.util.Set;
+import model.piece.Chariot;
 import org.junit.jupiter.api.Test;
 
 class BoardTest {
