+++ src/test/java/model/ChariotTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Map;
 import java.util.Set;
+import model.piece.Chariot;
 import org.junit.jupiter.api.Test;
 
 class ChariotTest {
