+++ src/test/java/model/GuardTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Map;
 import java.util.Set;
+import model.piece.Guard;
+import model.piece.Piece;
 import org.junit.jupiter.api.Test;
 
 class GuardTest {
