+++ src/test/java/model/SoldierTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Map;
 import java.util.Set;
+import model.piece.Piece;
+import model.piece.Soldier;
 import org.junit.jupiter.api.Test;
 
 class SoldierTest {
